pyproject.toml
[project]
name = "codefox"
version = "0.4.0"
description = "CodeFox CLI - code auditing and code review tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bm25s==0.3.0",
    "qdrant-client>=1.7.0",
    "fastembed>=0.3.0",
    "gitpython==3.1.46",
    "google-genai==1.63.0",
    "nltk==3.9.3",
    "numpy>=1.24.0",
    "ollama==0.6.1",
    "openai==2.21.0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "requests>=2.28.0",
    "rich==14.3.2",
    "typer==0.23.1",
    "tree-sitter-language-pack==0.13.0",
    "psutil==7.2.2",
    "PyGithub==2.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.8",
    "types-PyYAML>=6.0",
    "types-requests>=2.31.0"
]

[project.scripts]
codefox = "codefox.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
strict_optional = true
ignore_missing_imports = true

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.coverage.run]
source = ["codefox"]
branch = false

[tool.coverage.report]
exclude_lines = ["raise NotImplementedError", "if TYPE_CHECKING:"]
